refactor(storage): #71 2.1B-1 — extract DDL to storage/schema.py
Move every CREATE TABLE / CREATE INDEX / additive ALTER from
SQLiteEngineStore.init_schema (1,665 LOC) into a free function
apply_schema(connection) in storage/schema.py. The class-level
init_schema() shrinks to a 3-line wrapper that opens a commit-bracketed
connection and delegates.

_ensure_table_columns moves alongside as a module-private helper —
init_schema was its only caller (5 sites, all inside the moved body).

Schema text is byte-equivalent after whitespace normalization (verified
via diff of sqlite_master rows on a fresh engine.db before/after).
Tables: 74 → 74; indices: 66 → 66.

sqlite.py: 9,000 → 7,323 LOC.
